+++ GRADIO_PROJECT/.gitignore
@@ -1,0 +1,2 @@
+/.env
+/Clean_Train_Data.csv
