+++ GRADIO_PROJECT/flagged/log.csv
@@ -7,3 +7,6 @@ False,No,No,No,No,No Internet Service,No,No,No,No,No,No,Month-to-month,No,Bank t
 False,Yes,Yes,No,No,DSL,Yes,No,Yes,Yes,No,Yes,Month-to-month,Yes,Credit card (automatic),29,64,49-60 months,C:\Users\WORLDAX COMPUTERS\Python Projects\GRADIO_PROJECT\Embedding-a-Machine-Learning-Model-into-GUI-with-Gradio\flagged\output\tmpetfthb1w.json,,,2023-10-15 00:47:26.769352
 False,Yes,Yes,No,No,DSL,Yes,No,Yes,Yes,No,Yes,Month-to-month,Yes,Credit card (automatic),29,64,49-60 months,C:\Users\WORLDAX COMPUTERS\Python Projects\GRADIO_PROJECT\Embedding-a-Machine-Learning-Model-into-GUI-with-Gradio\flagged\output\tmp505fi6y2.json,,,2023-10-15 00:47:33.078650
 True,Yes,No,Yes,Yes,Fiber optic,Yes,Yes,,Yes,,Yes,Two year,Yes,Bank transfer (automatic),37-48 months,46,30,C:\Users\WORLDAX COMPUTERS\Python Projects\GRADIO_PROJECT\Embedding-a-Machine-Learning-Model-into-GUI-with-Gradio\flagged\output\tmpzwi7tsdm.json,,,2023-10-15 05:09:15.680928
+False,Yes,,Yes,,,,,,,,,,,,,15,15,C:\Users\WORLDAX COMPUTERS\Python Projects\GRADIO_PROJECT\flagged\output\tmpch9y7te9.json,,,2023-10-15 12:42:30.104358
+False,No,No,No,No,No Internet Service,No,No,No,No,No,No,Month-to-month,No,Bank transfer (automatic),0-6 months,20,150,Awaiting Submission...,Submit,,,2023-10-15 13:40:55.949385
+,,,,,,,,,,,,,,,,15,15,C:\Users\WORLDAX COMPUTERS\Python Projects\GRADIO_PROJECT\flagged\output\tmptd1f2uxp.json,,,2023-10-15 14:37:43.595917
