+++ GRADIO_PROJECT/flagged/output/tmptd1f2uxp.json
@@ -1,0 +1,1 @@
+{"label": "Prediction: Customer is likely to STAY", "confidences": [{"label": "Prediction: Customer is likely to STAY", "confidence": 1}, {"label": "Prediction: Customer is likely to LEAVE", "confidence": 0}]}
