+++ GRADIO_PROJECT/flagged/output/tmpzwi7tsdm.json
@@ -1,0 +1,1 @@
+{"label": "Prediction: Customer is likely to stay", "confidences": [{"label": "Prediction: Customer is likely to stay", "confidence": 1}, {"label": "prediction: Customer is likelt to LEAVE", "confidence": 0}]}
